+++ example/index.html
@@ -59,6 +59,8 @@
     </div>'
     var html = ejs.render(template, Browser)
     document.writeln(html)
+
+    console.log(Browser)
   </script>
 </body>
 </html>
